% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aam}
\alias{aam}
\title{Get metadata for objects in the Asian Art Museum of San Francisco.}
\usage{
aam(id, ...)
}
\arguments{
\item{id}{An object id}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata for objects in the Asian Art Museum of San Francisco.
}
\examples{
\donttest{
aam(17150)

aam(17140)
aam(17146)
aam(17144)
aam(17142)
aam(11462)
lapply(c(17150,17140,17144), aam)
}
}

