% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Get metadata for objects in the Art Institute of Chicago.}
\usage{
aic(id, ascii = FALSE, ...)
}
\arguments{
\item{id}{An object id}

\item{ascii}{(logical) Allow non-ascii characters. Set to \code{TRUE} to show
non-ascii characters. Default: FALSE}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata for objects in the Art Institute of Chicago.
}
\details{
Sorry, but the metadata returned in this function is a bit messy because the metadata
given for each piece is not stuctured, i.e., all elements are simply in html p tags, so its
hard to have consistent parsing.
}
\examples{
\donttest{
aic(40652)
aic(145241)
aic(87531)
aic(80538)
aic(11332)
aic(11287)
aic(210804)
aic(11475) # with artwork body description
aic(41033) # with history
lapply(c(11475,41594,59978,129410,41033,3394,21535,11475,41033), aic)
out <- aic(210765)
out
unclass(out)
browse(out)

aic(8254)
aic(25316)
}
}

