% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\alias{browse.aic}
\alias{browse.nga}
\title{Browse the object}
\usage{
browse(x, ...)

\method{browse}{aic}(x, ...)

\method{browse}{nga}(x, ...)
}
\arguments{
\item{x}{Output from any of the functions in this package.}

\item{...}{Further arguments passed on to \code{browseURL}}
}
\description{
Browse the object
}

