% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getty}
\alias{getty}
\title{Get metadata for objects in the Getty Museum.}
\usage{
getty(id, ...)
}
\arguments{
\item{id}{An object id}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata for objects in the Getty Museum.
}
\examples{
\donttest{
getty(140725)
getty(329471)
getty(138860)
getty(8197)
getty(3242)
getty(4967)
getty(112297)
getty(253190)
lapply(c(140725,8197), getty)

## doesn't work, different url and html schema
### see http://search.getty.edu/gri/records/griobject?objectid=301703057
getty(id=301703057)
}
}

