% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{met}
\alias{met}
\title{Get metadata for objects in the Metropolitan Museum of Art.}
\usage{
met(id, ascii = FALSE, ...)
}
\arguments{
\item{id}{An object id}

\item{ascii}{(logical) Allow non-ascii characters. Set to \code{TRUE} to show
non-ascii characters. Default: \code{FALSE}}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Get metadata for objects in the Metropolitan Museum of Art.
}
\examples{
\dontrun{
(out <- met(246562))
out$name
out$values
met(228901)
met(479283)
lapply(c(479283, 228901, 436876, 444244), met)
}
}
