% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moma.R
\name{moma}
\alias{moma}
\alias{moma_artists}
\alias{moma_artwork}
\title{Get metadata from Museum of Modern Art}
\usage{
moma_artists(...)

moma_artwork(...)
}
\arguments{
\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata from Museum of Modern Art
}
\examples{
\dontrun{
# artists
(arts <- moma_artists())

# artwork
(work <- moma_artwork(config = verbose()))
}
}
