% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musemeta.R
\docType{package}
\name{musemeta-package}
\alias{musemeta}
\alias{musemeta-package}
\title{R client for museum metadata}
\description{
R client for museum metadata
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

