% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nga.R
\name{nga}
\alias{nga}
\title{Get metadata for objects in the National Gallery of Art.}
\usage{
nga(id, ascii = FALSE, ...)
}
\arguments{
\item{id}{An object id}

\item{ascii}{(logical) Allow non-ascii characters. Set to \code{TRUE} to show
non-ascii characters. Default: FALSE}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata for objects in the National Gallery of Art.
}
\examples{
\dontrun{
nga(id=33267)
nga(47242)
nga(47243)
nga(45987)
nga(33268)
lapply(c(75861,20820,75959,27809,143679,27773,28487), nga)

library('httr')
nga(45987, config=verbose())

browse(nga(33268))
}
}
