% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scrapi}
\alias{scrapi}
\alias{scrapi_info}
\alias{scrapi_random}
\alias{scrapi_search}
\title{Use the scrapi API for The Metropolitan Museum of Art data}
\usage{
scrapi_random(fields = NULL, ...)

scrapi_info(id, fields = NULL, ...)

scrapi_search(query, ...)
}
\arguments{
\item{fields}{(character) One or more fields to return in a vector}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}

\item{id}{(numeric) An object id}

\item{query}{(character) Query terms}
}
\description{
Use the scrapi API for The Metropolitan Museum of Art data
}
\examples{
\donttest{
# Get a random object
scrapi_random()
scrapi_random(fields=c('medium','whoList'))

# Get object information
scrapi_info(123)
scrapi_info(123, fields='title')
scrapi_info(123, fields=c('title','primaryArtistNameOnly','medium'))

# Search for objects
scrapi_search(query='mirror')
scrapi_search(query='siphon nozzle')
}
}

