% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aam.R
\name{aam}
\alias{aam}
\title{Get metadata for objects in the Asian Art Museum of San Francisco.}
\usage{
aam(id, ascii = FALSE, ...)
}
\arguments{
\item{id}{An object id}

\item{ascii}{(logical) Allow non-ascii characters. Set to \code{TRUE} to show
non-ascii characters. Default: FALSE}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata for objects in the Asian Art Museum of San Francisco.
}
\examples{
\dontrun{
aam(17150)

aam(17140)
aam(17146)
aam(17144)
aam(17142)
aam(11462)
lapply(c(17150,17140,17144), aam)
}
}

