% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getty.R
\name{getty}
\alias{getty}
\title{Get metadata for objects in the Getty Museum.}
\usage{
getty(id, ascii = FALSE, ...)
}
\arguments{
\item{id}{An object id}

\item{ascii}{(logical) Allow non-ascii characters. Set to \code{TRUE} to show
non-ascii characters. Default: \code{FALSE}}

\item{...}{Curl args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Get metadata for objects in the Getty Museum.
}
\examples{
\dontrun{
getty(233750)
getty(138860)
getty(8197)
getty(3242)
getty(4967)
getty(112297)
lapply(c(140725,8197), getty)
}
}
