% Generated by roxygen2 (4.0.2): do not edit by hand
\name{met}
\alias{met}
\alias{met_get}
\alias{muse_get}
\title{Get metadata for an object id.}
\usage{
met_get(id, ...)

muse_get(id, ...)
}
\arguments{
\item{id}{An object id}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata for an object id.
}
\details{
This function currently only gets data for The Metropolitan Museum of Art

\code{muse_get} has changed to \code{met_get}
}
\examples{
\donttest{
out <- met_get(559490)
out$name
out$values
met_get(246562)
lapply(c(479283, 228901, 436876, 444244), muse_get)
}
}

