% Generated by roxygen2 (4.0.2): do not edit by hand
\name{muse_get}
\alias{muse_get}
\title{Get metadata for an object id.}
\usage{
muse_get(id, ...)
}
\arguments{
\item{id}{An object id}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
Get metadata for an object id.
}
\examples{
\donttest{
muse_get(559490)
muse_get(559490)$name
muse_get(559490)$values
muse_get(246562)
}
}

