% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapi.R
\name{scrapi}
\alias{scrapi}
\alias{scrapi_get}
\alias{scrapi_info}
\alias{scrapi_random}
\alias{scrapi_search}
\title{Use the scrapi API for The Metropolitan Museum of Art data.}
\usage{
scrapi_random(fields = NULL, ...)

scrapi_info(id, fields = NULL, ...)

scrapi_search(query, ...)

scrapi_get(x, ...)
}
\arguments{
\item{fields}{(character) One or more fields to return in a vector}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}

\item{id}{(numeric) An object id}

\item{query}{(character) Query terms}

\item{x}{URL or ID for a scrapi object.}
}
\description{
NOTE: THE RANDOM ENDPOINT IS DOWN FOR NOW.
}
\details{
You can also use the \code{\link{met}} function. The equivalent function with the
scrapi API is \code{\link{scrapi_get}}. The latter gets much more data, and uses a REST API,
while the former scrapes the html directly.
}
\examples{
\dontrun{
# Get object information
scrapi_info(123)
scrapi_info(123, fields='title')
scrapi_info(123, fields=c('title','primaryArtistNameOnly','medium'))

# Search for objects
scrapi_search(query='mirror')
scrapi_search(query='siphon nozzle')
scrapi_search(query='22.1.962')

# Get an object
## with a url for a scrapi object
scrapi_get("http://scrapi.org/object/427581")
## with an object id
scrapi_get(427581)

# RANDOM down for now
## scrapi_random()
## scrapi_random(fields=c('medium','whoList'))
}
}
\references{
http://scrapi.org \url{https://github.com/jedahan/collections-api}
}
\seealso{
\code{\link{met}}
}

